+++ packages/Core/src/adapters/snapDB.ts
@@ -1,5 +1,5 @@
 import { InanoSQLAdapter, InanoSQLDataModel, InanoSQLTable, InanoSQLPlugin, InanoSQLInstance, VERSION, configFilter } from "../interfaces";
-import { noop, deepFreeze, generateID, binarySearch, assign, cast, blankTableDefinition, deepSet, chainAsync } from "../utilities";
+import { noop, deepFreeze, generateID, binarySearch, assign, cast, blankTableDefinition, deepSet, chainAsync, deepGet } from "../utilities";
 import { nanoSQLMemoryIndex } from "./memoryIndex";
 import { SnapDB } from "snap-db";
 import * as fs from "fs";
@@ -44,7 +44,13 @@ export class SnapDBAdapter extends nanoSQLMemoryIndex {
 
     _path: string = typeof process !== "undefined" ? process.cwd() : "";
 
-    constructor() {
+    constructor(public snapDBArgs?: {
+        dir: string;
+        key: "string" | "float" | "int";
+        cache?: boolean;
+        autoFlush?: number | boolean;
+        mainThread?: boolean;
+    }) {
         super(true, false);
         this._ai = {};
         this._tableConfigs = {};
@@ -69,7 +75,11 @@ export class SnapDBAdapter extends nanoSQLMemoryIndex {
     createTable(tableName: string, tableData: InanoSQLTable, complete: () => void, error: (err: any) => void) {
         this._tableConfigs[tableName] = tableData;
 
-        this._tables[tableName] = new SnapDB({dir: path.join(this._baseFolder, tableName), key: tableData.isPkNum ? "float" : "string"});
+        this._tables[tableName] = new SnapDB({
+            dir: path.join(this._baseFolder, tableName), 
+            key: tableData.isPkNum ? "float" : "string",
+            ...this.snapDBArgs ? this.snapDBArgs : {}
+        });
 
         if (this._tableConfigs[tableName].ai) {
             this._tables["_ai_store"].get(tableName).then((aiValue) => {
@@ -82,6 +92,20 @@ export class SnapDBAdapter extends nanoSQLMemoryIndex {
         } else {
             this._tables[tableName].ready().then(complete).catch(error);
         }
+    }
+
+    batch(table: string, actions: {type: "put"|"del", data: any}[], success: (result: any[]) => void, error: (msg: any) => void) {
+        this._tables[table].begin_transaction().then(() => {
+            return Promise.all(actions.map(a => {
+                if (a.type === "put") {
+                    return this._tables[table].put(deepGet(this._tableConfigs[table].pkCol, a.data), JSON.stringify(a.data));
+                } else {
+                    return this._tables[table].delete(a.data);
+                }
+            }))
+        }).then(() => {
+            return this._tables[table].end_transaction();
+        }).then(success).catch(error);
     }
 
     dropTable(tableName: string, complete: () => void, error: (err: any) => void) {